+++ src/Mod/Sketcher/App/CMakeLists.txt
@@ -79,6 +79,8 @@ SET(Datatypes_SRCS
     GeometryFacade.h
     ExternalGeometryFacade.cpp
     ExternalGeometryFacade.h
+    GeoEnum.cpp
+    GeoEnum.h
 )
 SOURCE_GROUP("Datatypes" FILES ${Datatypes_SRCS})
